+++ packages/native-with-source/src/deps/PDFWriter/InputPFBDecodeStream.cpp
@@ -586,7 +586,7 @@ EStatusCode InputPFBDecodeStream::ReadDecodedByte(Byte& outByte)
 Byte InputPFBDecodeStream::DecodeByte(Byte inByteToDecode)
 {
 	Byte result = (Byte)(inByteToDecode ^ (mRandomizer >> 8));
-	mRandomizer = (unsigned short)(((inByteToDecode + mRandomizer)* CONSTANT_1 + CONSTANT_2) % RANDOMIZER_MODULU_VAL);
+	mRandomizer = (unsigned short)(((unsigned int)(inByteToDecode + mRandomizer) * CONSTANT_1 + CONSTANT_2) % RANDOMIZER_MODULU_VAL);
 	return result;
 }
